functions/process_sqs_queue/requirements.txt
aws-psycopg2==1.3.8
aws-lambda-powertools==2.19.0
boto3==1.28.1